canais/requirements.txt
Flask
requests
python-dotenv
PyJWT
flask-cors
